app.py
import streamlit as st
import os

import torch
from transformers import AutoTokenizer, AutoModelForSequenceClassification
from dotenv import load_dotenv
#from google-generativeai import generate

# Load environment variables (Optional)
load_dotenv()

## Function to load DistilBERT model and classify sentiment
MODEL_NAME = "distilbert-base-uncased-finetuned-sst-2-english"  # Sentiment analysis model

# Define the model outside the function (before calling analyze_sentiment)
model = AutoModelForSequenceClassification.from_pretrained(MODEL_NAME)

def analyze_sentiment(news_text):
  """
  This function analyzes the sentiment of the provided news text using a DistilBERT model.
  """
  # Preprocess the text using the tokenizer
  tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME)
  inputs = tokenizer(news_text, return_tensors="pt")

  # Get the model's predictions
  outputs = model(**inputs)
  predictions = torch.argmax(outputs.logits, dim=-1)

  # Retrieve the predicted sentiment label based on the model's output format
  # (Note: This might need adjustment depending on the specific model's output)
  sentiment_labels = ["NEGATIVE", "POSITIVE"]  # Adjust labels if necessary
  predicted_sentiment = sentiment_labels[predictions.item()]

  return predicted_sentiment

# Streamlit app section
# User Input for News Analysis
news_text = st.text_area("Enter News Text Here:", placeholder="Paste news article here")

if news_text:
  # Define tokenizer here (already defined before)
  sentiment = analyze_sentiment(news_text)
  st.write(f"Sentiment Analysis: {sentiment}")

# User Input for What-If Scenario
scenario_text = st.text_input("Enter your 'What-If' scenario here:", placeholder="e.g., What if interest rates rise?")

if scenario_text:
    # Build the prompt string
    prompt = ""

    # Call Gemini AI (replace with your implementation)
    # response = generate(prompt=prompt)  # Assuming you have a 'generate' function

    # Process the response (replace with your implementation)
    # generated_text = extract_text_from_response(response)  # Assuming you have an 'extract_text_from_response' function

    # Display the results
    st.write("**AI Scenario Analysis:**")
    #st.write(generated_text)  # Replace with the actual generated text

# Downloading model files warning (can be ignored)

