requirements.txt

transformers
streamlit
google-generativeai
python-dotenv
torch
 